Refactor navigation components into separate files
- Moved primary and secondary navigation links to constants for better organization.
- Created PrimaryNav and SecondaryNav components to handle respective navigation logic.
- Updated TopNav to utilize new components and manage secondary navigation state.
- Improved code structure and maintainability.
