+++ src/components/Navigation/constants.ts
@@ -1,0 +1,21 @@
+import type { NavLink } from "./types";
+
+export const primaryNavLinks: NavLink[] = [
+  { href: "/users", text: "Users" },
+  {
+    href: "/services",
+    text: "Services",
+    children: [
+      { href: "/services/kliento", text: "Kliento" },
+      { href: "/services/servers", text: "Services" },
+    ],
+  },
+  { href: "/overview", text: "Tech overview" },
+  { href: "/about", text: "About" },
+];
+
+export const secondaryNavLinks: NavLink[] = [
+  { href: "/services/kliento", text: "Clients", id: "clients" },
+  { href: "/services/kliento/servers", text: "Servers", id: "servers" },
+  { href: "/services/kliento/services", text: "Overview", id: "overview" },
+];
