+++ src/components/Navigation/index.ts
@@ -1,0 +1,3 @@
+export { default as TopNav } from "./TopNav";
+export { PrimaryNav } from "./PrimaryNav";
+export { SecondaryNav } from "./SecondaryNav";
